Tim Proescholdt :
For the moment I have only one change, to eliminate a bug when someone
harvests a OAI set whose name contains a "-".

In ResumptionToken.java

Line 40:

-public static final String SEPARATOR = "-";
+public static final String SEPARATOR = "/";

The reason for this is that there can be problems with OAI "sets" (AKA
categories having "-" in the name. "/" is unusual for a name, "-" is not.


git-svn-id: https://geonetwork.svn.sourceforge.net/svnroot/geonetwork/branches/2.6.x@6690 ff403467-3f20-0410-99e5-ff581e626a08