added a new profile 'Monitor'.
A 'Monitor' is a user with the Monitor profile which is designed to allow a user that has access to monitoring but no edit capabilities.
The logic is essentially: some monitoring can be expensive and a good vector for denial of service attacks therefore they need to
be secured.  With this new Profile (Administrator extends this profile) we can have a user that is specifically designed to be able to
monitor but not other activities.  It is not in the normal straight profile hierarchy so Editor and Reviewers (etc...) do not have monitor
capabilities

git-svn-id: https://geonetwork.svn.sourceforge.net/svnroot/geonetwork/trunk@9181 ff403467-3f20-0410-99e5-ff581e626a08