Added a new profile 'Monitor'
A 'Monitor' is a user with the Monitor profile which is designed to allow a user
The logic is essentially: some monitoring can be expensive and a good vector for
be secured.  With this new Profile (Administrator extends this profile) we can h
monitor but not other activities.  It is not in the normal straight profile hier
capabilities

git-svn-id: https://geonetwork.svn.sourceforge.net/svnroot/geonetwork/branches/2.8.x@9182 ff403467-3f20-0410-99e5-ff581e626a08