Fix for a potential short term memory leak
Since LuceneSearcher is often cached in the Session a Spatial Filter might also be cached.  Since the Spatial Filter needs access to the Spatial index it follows that the spatial index will also be cached.  In large catalogs this can be several MB.  In once case with only 12,000 records this was 8MB.  Apparently it was being updated regularly because there were several UserSessions with different index instances thus all the usersessions were taking up nearly 30 MB.