Improve MyOcean view.
Including GeoNetwork tickets:
 * Editor / gco:ScopedName and LocalName always set to Name
Labels for those 2 elements does not change according to GUI languages and could not be overriden in a profile.

 * Editor / Add the option to overrides gml:time* or gml:duration element title
Add the title parameter in order to force a custom title for those elements.
This could be useful to replace "timeposition" by "begin" when making a view in flat mode accessing directly the element and not its parent (to build a full section).

 * Editor / Add a function to create a add subtemplate button outside an element control
Currently add subtemplate button is only available in the last element of a type. Add a generic function to create a custom button to be place outside element scope (eg. at the head of a section).