check file timestamp to see if favicon and dummy image need to be loaded
The favicon was cached which broke the logo management functionality.  Now the files like favicon and dummy image are checked to see if a newer version is on disk.
If the timestamp on the file indicates that it is newer than the version in the cache then file is loaded again.