MyOcean view / End date improvements (#14459).
The end date can be unknown. Thus for this input a checkbox named "unknown" and a date input (both inputs on one line) are available.
If the checkbox is checked then the date input becomes disabled. If not then the date input is enabled and a date must be selected by the user.