added default namespace for each thesaurus to make editing thesauri easier
Previously when one added an element to a thesaurus they would add a keyword and then they needed to change the id to add a namespace.
If they didn't change namespace then when the thesaurus was restarted the namespace would be a file uri to the thesaurus file.  This
is a problem if the id was used anywhere for identifying the keyword.

Fixes: http://trac.osgeo.org/geonetwork/ticket/1190