Implemented more efficient and flexible way of language determination on first contact.
This change modifies the welcome page, login page and access denied services.  Previously these pages would be jsps which execute javascript on the client side which detects the current language and redirects the user based on the language to the language specific srv/ service.
This changes the behaviour in several ways:

* The language is detected on the server side based on:
  * If an 'hl' request parameter is specified then it will be used
  * Next in the 'geonetwork_Preferred_Language_Cookie' cookie is checked for and used (the cookie is now set in JeevesServlet)
  * Next the 'Accept-Language' header is used if it is present (which is normally the case)
  * Next the language specified in config.xml is used
  * Finally the Geonet.DEFAULT_LANGUAGE is used (typically eng)

As aluded to above, a cookie is set in JeevesServlet so that when the user changes his/her
gui language the prefered language for that user is set (via cookies) so that that that language
will be used in the future for this calculation (until the language is changed or the cookie expires
after its limit of 1 week.