Restrict resource filtering on web-client and web-ui so that only the specific files that need filtering are filtered.
In web-client especially there are many cases where ${...} is used in javascript files but are not intended to be replaced by maven build because they are used for string formatting.  If the build machine has one of those strings as
a system property or environment variable it will be replaced resulting in javascript that is incorrect.