Add WMS thumbnail maker panel to editor.
This panel is added to the popup which allows to add thumbnail by URL or Upload.
A new mode is added providing a map with the background layer and all WMS layers registered
in the metadata record. User can choose a layout, a scale and an area of interest
in order to print the map in PNG format and attach it to the current metadata record.

Changes:
* Update MapFish print to 2.0 version
* Move MapFish print module configuration and resources file to WEB-INF/config-print (instead of webapp root folder)
* Add layout and PDF template for printing thumbnail