LDAP mapping - highest profile election (mantis #21684)
In case of many listesiteweb from the LDAP, we were taking the first one
as profile for a given group ; hence, if a user had the following
attributes:

listesiteweb: SXT5_AIRESMARINES_RegisteredUser
listesiteweb: SXT5_AIRESMARINES_Reviewer

The first element returned by the list in the code was used (the order
should not be considered as predictable). We need to elect the highest
profile at a group scope, instead of taking the first one.

Manually tested, the current testsuite for this specific class has not
be broken, but no automatic test has been added for this specific case.