Harvester / Category not saved #578
This is more a workaround than a fix but fix settings categories
on first harvest and updates.

The issue is related to some method saving the metadata object
to the repository and caller method also using a previous version of
the metadata object. The fix retreive an updated version of the
metadata object when needed in order to properly save the record.