Add IndexingTask which will be trigger every 30sec by default to reindex a set of records.
This set of records is maintained in the IndexingList bean which can be populate whenever a record needs to be indexed in the near future (not immediate indexing). For example, this is the case when a record is viewed and the popularity updated. If a user switch from one tab to another for the same record, the popularity is updated in the database for each tab switch but the record will be reindexed only once every 30sec (instead of each time in previous IncreasePopularityTask).