Fix NPE when LDAP user login (#708).
Do not reset profile using LDAP info, if configuration says to use the local db value.
Do not reset user information if LDAP info are empty (and not only if null). With this, LDAP user info could still be set in the database even if not available in the LDAP.

Try to add doc to ldap.privilege.create.nonexisting.users property.