Search on keyword code depending on type of search parameter (eg. equals, contains, startswith).
Actually when running a keyword search, both the labels and the code are used for the search.
The label search is based on the search type parameter but the code search is equal only.

Use the search type parameter to search in keyword codes like on labels.

This will allows user to search keyword by code without knowning the full URI
eg. http://www.eionet.europa.eu/gemet/concept/12349 should be used for now to search
for "beach cleansing" concept in GEMET. With this change 12349 could also be used.

This is useful when codes may have meanings or when users know concept codes.