Document how to workaround the hibernate issue on schema.
If you have 2 GeoNetwork pointing to the same db and different schema,
hibernate does not create the sequence in the correct place.

This commit proposes a workaround on the schema management
by Hibernate which is kind of broken in certain
circumstances. It will force the hibernate sequence to be created into
the expected schema (and not the public one).

Tests: Tested with 2 different tomcats / geonetwork configured to
connect onto the same database, with the following configuration:

* Geonetwork "sextant-sdn": db bug_sdn, schema bugsdn, user/pass
bugsdn/bugsdn, default search_path: bugsdn.
* Classic geonetwork: db bug_sdn, schema bugsdn2, user/pass
bugsdn2/bugsdn2, default search_path: bugsdn2.

Launching the 2 tomcats with empty schemas will populate both bugsdn and
bugsdn2, with the schema explicitely set into
config-spring-geonetwork.xml:

<entry key="hibernate.default_schema" value="bugsdn"/>
and
<entry key="hibernate.default_schema" value="bugsdn2"/>
respectively.

Thanks @pmauduit for investigating and suggesting this fix.