Change templates to use the _logo index field for logo path.
_logo is chosen as follows:

If metadata.groupOwner != null && metadata.group.logo != null && settings.prefergrouplogo {
  _logo = /resources/logo/{{metadata.group.logo}}.<ext>
} else {
  _logo = /resources/logo/{{metadata.source.uuid}}.<ext>
}

<ext> is chosen by searching the resources/logo directory for an image with the prefix source.uuid or group.logo (according to above algorithm).  This ensures that the image is valid irregardless of whether the user uploaded a gif or png or tif or... and is not hardcoded as a .gif like many old solutions of logo were.

This will not change any behaviour of existing systems because most current sytems will not have group.logo since that is a new geocat 3.0 feature.  additionally there is a setting that will force the logo to always be based on the metadata source.