Local FS harvesting - fixing memory consumption
When the harvester is configured to delete old previously harvested
elements, a query is made to fetch *every* (java object) Metadata for
this specific harvester, from the database.

On regular directories, containing less than a hundred or thousands of
MD, that is not a problem, but our customer is willing to harvest a
directory containing +230k metadata, which obviously won't fit even on a
server with a huge amount of RAM.

Most problematic line of code is here:
https://github.com/geonetwork/core-geonetwork/blob/develop/harvesters/src/main/java/org/fao/geonet/kernel/harvest/harvester/localfilesystem/LocalFilesystemHarvester.java#L133

This commit proposes to address the problem by:

- Constructing a list of Integers instead of strings (this can reduce
  the amount of java objects created on the heap by a factor between
  1.4 and 4.4, from a little test which compares constructing huge
  lists of List<String> vs List<Integer>) ;

- Getting only a List<Integer> instead of a List<Metadata>, when
  checking which MDs have to be deleted.

Tests: runtime tested, on a tomcat with -Xmx2048M
Test suite: fails, but on the following one:
  InspireAtomFeedRepositoryTest.testCleanAtomDocumentsByMetadataId: DataIntegrityViolation
So I'm not sure this commit introduced this as a regression.