Fixed xpath with position so that position.
Previously you the position was calculated as the position of elements rather than the position of the elements with the given name.
For example consider

    <metadata>
      <identification>
        <child>...
        <otherchild>...
        <child>  <!-- desired element-->
      </identification>
    </metadata>

Previously the xpath would be `identification/child[3]` instead of `identification/child[2]` as it should be. If the position was [3] then the xpath should have been: `identification/node()[3]` not `identification/child[3]`