ISO19139 / INSPIRE Conformity
Improve INSPIRE form to follow the encoding rules defined
in MD_IR_and_ISO_20131029.

Conformity can be:

* true if conformant
* false if not conformant
* null (with nilReason = “unknown”) if not evaluated

It used to be not evaluated if not defined in the XML document.

A custom directive is created in order to handle 3 states checkbox
populating a nilReason attribute with the third value defined.