formatters - fixing directory traversal / file disclosure vuln
The current code for the file edition services, which is used to update
files in the formatter interface allows an admin to query whatever files
from the server filesystem, querying an absolute file in the parameter
(see test).

The vulnerability is mitigated since the service shall only be
accessible to user admins or higher profiles.