User session warning.
After identification, a user session is created. This session will be closed automatically at some point by the server for security reasons. Default session timeout is set to 35 min. Messages are displayed when the session is about to expire:

* When there is no activity in the browser and that the session is about to expire (3 min before), a warning is displayed next to the user details.
* 1 min before the end of the timeout, another message is displayed to recall to the user that pending work should be saved.
* When the session has been probably cancelled by the catalog, a message recommend to refresh the page and sign in again if needed.

Session last activity and expire time are added to the cookie by a new filter.