LDAPUser - fixing potential NPE
In some specific spring security cases, the object is compared to NULL,
making the login process fail. This patch allowed to be able to log in
onto a georchestra instance.

TODO: I could not manage to find why this was needed in details, but the
modification seems harmless to me. If anyone has a better understanding
please do comment.