fixing images logo urls
concatenating a date could result in strange characters added in the
URL, using a timestamp here sounds more relevant. In addition the
concatenated string was containing extra double-quotes after
substitution, making the image loading fail because of invalid URL.

In addition I was advised to replace the src= by data-ng-src to be
coherent, but since I'm not confident myself with AngularJS, I cannot
tell.