Resolving environment variables in overrides
In geOrchestra, we are willing to stay the most close to the GN
codebase, as a result, we use the overrides mechanism to resolve
configuration customization files. But one limitation of this feature is
that using java properties is not possible in the XML file. We are then
forced to hardcode the paths in the provided XML overrides, see:

https://github.com/pmauduit/core-geonetwork/blob/georchestra-gn3-3.0.x/web/src/deb/resources/etc/georchestra/geonetwork/config/config-overrides-georchestra.xml

In our case, we generate a debian package for GeoNetwork, so we can
guess that the files (for what we called the "georchestra datadir",
hosting our configuration files), would end up in the right place (e.g.
/etc/georchestra/geonetwork), but an administrator could define his own
datadir elsewhere, forcing him to update the overrides file.

Hence, the purpose of this PR allows the use of ${env:prop} variables to
load the extra XML configuration files.

Tests: unit tests added, untested at runtime for now.

Drawbacks: The current code could be enhanced to ensure that the
${env:prop} are resolved elsewhere, but I am considering this as a good
start. Feedback of the community welcome.