Introducing the new IMetadata superclass.
This means a lot of small changes in several places where metadata is used.
This also means a big change on how GeonetEntity creates xml (asXml).

All the tests pass. Not warranteed that all the functionalities remain stable, because now there are database searches that have to handle two tables at the same time. Lot of room for improvement.