Directory / Add service to collect and synchronize entries.
2 services are added:
* GET|PUT /srv/api/0.1/registries/actions/entries/collect
* GET|PUT /srv/api/0.1/registries/actions/entries/synchronize

Collect scan one or more records for element matching the XPath provided and save them as directory entries (ie. subtemplate).

Synchronize scan one or more records for element matching the XPath provided and then check if this element is available in the directory. If Found, the element from the directory update the element in the record and optionally properties are preserved (eg. contact role). The identifier XPath is used to find a match. If not, the UUID is based on the content of the snippet. It is recommended to use an identifier for better matching (eg. ISO19139 contact with different roles will not match on the automatic UUID mode).

Collect service used to be available in old XSL user interface but was never ported to ExtJS interface. It was made by @sppigot and was named ```metadata.batch.extract.subtemplates```. It was also improved:
* UUID are computed from a property identified by an XPath (eg. email) and if not set, from the XML snippet itself (only the last mode was available).
* Past service was also substituting if requested using XPath. Now substitution is made in the synchronize service and could be made using XLink or not

Other related changes:
* ISO19139 / Subtemplate indexing / Add email field (used for searching entries by email)
* Spring services / Add a DOMElementMessageConverter to return Element in Spring MVC services
* Remove old service.