Editor / Allow editing of keyword in multilingual metadata records.
GUI language MUST match one of the metadata languages.

Keyword in gui language is used as reference to init editor.

Presence of this keyword is then checked in corresponding thesaurus.
If present, directive is initialized.
If not, keyword is preserved and warning displayed.

This change also add support to metadata having locale identifiers
not based on iso2letter code. Language identifiers are sent to
thesaurus service creating keyword fragments.