Fix user unable to login after password reset in the admin console
If a UserAdmin user reset the password of and old user with an old
password hash (SHA-1) he is not able to login because update password
set the hash using SHA-256, but the column `security` still contains
`UPDATE_HASH_REQUIRED`.
This commit removes that string from `security` after updating the
password.
It also fixes some errors reported by Findbugs.