Selection / Add multiple selection support.
Selection used to be using one session property Geonet.Session.SEARCH_RESULT
used by all services. This change introduces an extra parameter ```bucket```
which can be used to create multiple selections in the same app. eg. the
selection made in the main search page does not interact with the batch process
selection.

On the client side, all JS apps using selection defines a unique bucket name
to not interact with each others. If not set, a unique bucket identifier is set.