Don't use proxy for URL with the same origin (#1943)
CORS errors and network/server errors can't be distinguish by the browser JS engine.
CORS errors usually happen when the requested URL is in other domain/protocol that
the current page current one.
With this commit request made to the same origin that GN that result in an error will
not be retried using the proxy and the request will be rejected with the actual response
value from the server.
If the request is for another origin it will be retried using the proxy.