Add concept of template of subtemplate.
Currently there is no capability to create a directory entry from
a template. It has to be imported as XML. A template of subtemplate
provides the possibility to define which elements should be part
of directory entry (eg. a contact for INSPIRE define an organisation name
an email and a role).

Template of subtemplate are identified by "sub-tpl-" prefix in
the template directory.

It will allow to provide to user the capability to create new entries
when none found in the directory.