Improve directory permissions management
Now the permission widget is cleared when not displayed, which prevents some bugs & inconsistencies.
Also a slight improvement to the directory UI search results.

To do this I had to handle a `closeCallback` parameter on `gnModal`, which is pretty handy for this sort of case (ie doing something when the modal is closed by any means).