Editor configuration / Add custom sidepanel support.
Right side panel was partially customizable using the admin configuration which depends on schema.
Therefor, when plugin a new schema, this setting requires manual change in the JSON settings.

This commit cover the following improvements:
* sidepanel can be configured in config-editor.xml
* sidepanel can be different in each views (or totally removed)
* sidepanel order of widgets is defined in the configuration
* sidepanel can contains text block with instructions
* editor can contain text element (not included in a section eg. for a view header)
* editor can contain directives (like the one use in the sidepanel)
* online source list directive can now cover a restricted set of relations. This
allows to have one directive covering thumbnails in the editor directly and another
one in the sidepanel covering other types.
* editor help can be display as tooltip hover icons or on focus on fields