Use standard properties for proxy
After the introduction of the new ProxyServlet it doesn't use the proxy
gateway configured in the admin settings making impossible the connection
to external websites in environments where a proxy is needed to access
the external network.
This commit allows ProxyServlet to use the proxy configured by Java
standard properties http.proxyHost, http.ProxyPort, etc.