Fix error when printing WMTS layers using REST
Since GN uses Mapfish-print v2.1.2 and it contains some bugs when
printing a WMTS layer when REST is used this commit is a workaround that
processes some parameters in the client side to allow Mapfish print the
layer.
For example it needs a `dimensions` parameter, even when it should be
optional. Another change is that Mapfish v2.1.2 doesn't replace the
`{style}` and `{layer}` parameters in the URL, so we need to do that
in the Javascript code.