Don't allow to create more duplicates usernames using the API
Checks if there are already other users using the same username with different
letter case when creating a new user.
When updating a user only allow set a duplicate username if the user already had
this username with different case and the new name doesn't collide with any existing
username.
Issue #2464