Baselayer not changing sometimes
Sometimes after changing the base layer selected the layer displayed in the map doesn't change.
This is because the layer is not added to the map in the right position. To solve this first remove all
the background layers in the map and then add only the one selected in the base layer switcher.