Create users from LDAP in database using lowercase name
This commit introduces a new LDAP configuration property,
`ldap.usernameCaseInsensitive` to indicate that LDAP search are case
insensitive. In this case, the user created in the database will be
lowercase, independently of how the user enters the username in the
login input field.

If the new property `ldap.usernameCaseInsensitive` is explicitly set to
false then the user created in the database will have the exact same
case the user entered in the login box, allowing to exist multiple users
with the same username but different case combinations, for example
`user1`, `User1` and `uSEr1`.