Formatter / XSLT / Display temporal extent properly based on config editor #2868
Currently a CHOICE_ELEMENT element is added when building the metadocument for the editor. The XPath for matching a gml:beginPos or gml:endPos required an extra * parent to match due to this extra element added in between. BUT the XSLT formatter using the metadata document was failing to find a match due to that. Use // instead to match in both cases.

This is affecting custom views using a template field for temporal extent.