Access rights overview panel.
This new panel only accessible by administrator user when ElasticSearch index is populated
provides an overview of privileges for a set of records.

This panel can be accessed from the editor board.

Records can be filtered by title and publication group.
User can choose which types of operation and groups to preview.

This is using scripted fields in ES to create a field easily from a
simple expression; here, does operation view is available for group A?