Don't serialize the full list of category's records to XML/JSON
Fixes a problem of `OutOfMemoryError` when the catalog has a lot of records with
categories and the list of Groups is requested using the service `admin.group.list`
since all the records asigned to each category was being returned to as part of
the `records` collection.
This commit excludes the records property from the XML or JSON generated.