[BP] Throws a more legible exception when region is not found (#3069)
Backport of 2da2ef5.

When a spatial search uses the `region:` prefix and no region is found
for it in the thesaurus throw an `IllegalArgumentException` with message
"Geometry %s not found in the available source RegionDAO objects"
instead of letting JTS try to parse a null geometry and throw the
cryptic "Argument "env" should not be null." error.

Also re-indent the class, organize the imports and upgrade the use of
StringUtils class from Apache Commons to v3.