[BP] OutOfMemoryError when listing groups using service admin.group.list (#3079)
Backport of 9af2f2cfaee98ec5a14dd284dffea740e55975ed.

* Don't serialize the full list of category's records to XML/JSON

Fixes a problem of `OutOfMemoryError` when the catalog has a lot of records with
categories and the list of Groups is requested using the service `admin.group.list`
since all the records asigned to each category was being returned to as part of
the `records` collection.
This commit excludes the records property from the XML or JSON generated.

* Retrieve the category's record count using count

Retrieve the number of records asigned to a category using a `count` query
instead of loading all the records and counting the size of the collection
returned.