LDAP / Preserve user privileges on sign in.
When a LDAP configuration define a default privilege for a group
eg.
```
ldapUserContextMapper.mapping[privilege]=privileges,eea_users
```
then on sign in, the user lost any other privileges which might
have been assigned from the admin console and defaults are restored.

If an LDAP user has privileges defined, preserve them and do not
apply the default one. If none, add the default.