API / Formatter / Add JSON and text responses support.
Some formatter may provide text responses and not only XML or PDF as
currently supported. Rename file with a proper extension based on file
type. Add support to formatter providing a text response. In such case
the XSLT formatter has to return one element like the following with the
text to be returned:

```
  <xsl:template match="/">
    <textResponse>
    {
    "@context": "http://schema.org",
    "@type": "Dataset",
    "@id": "https://doi.org/10.5438/4K3M-NYVG"

     <!-- <xsl:apply-templates select="$metadata"
                           mode="toJsonLd"/>-->
    }
    </textResponse>
```

The text can be CSV, or JSON, ...

The content type can still be controlled using accept header or the
output URL parameter.

eg.
http://localhost:8080/geonetwork/srv/api/records/ff8d8cd6-c753-4581-99a3-af23fe4c996b/formatters/jsonld?output=json