UI configuration improvement.
The ui/config settings introduced
https://github.com/geonetwork/core-geonetwork/pull/1848 (and inspired
from Sextant) is somehow a bit hard to maintain.

The default value is now set from CatController only.

The default db value is empty to use the one from CatController until a
db value is defined by the catalogue admin.

JsonDatabaseMigration task should be used to update the settings (see
https://github.com/geonetwork/core-geonetwork/pull/2808). We should not
set the full value in SQL migration script anymore (BTW this is a line
in conflict in most of the PR since then).

JSON editor is now used instead of the textarea.

Future work required? Maybe we should find a better way to merge the db
value and update with the latest.