Fixes harvester type names not translated (#3362)
Due to timing issues while loading the Harvester admin page and the language files
sometimes the translation for the harvester type is not available yet when the page
is rendered. This is cause by the use of `$translate.instant` function. To fix this
and allow the translation to be set once the language is available we use now a promise
for setting the `text` translated value.