Map viewer not zooming to context extent
When a record describes a map and a context is associated to the record, when loading the map from the record view the map may not zoom to the map extent defined in the context. When loading from the map panel, it works. So it looks like it is related to the fact that the map may not be initialized early enough in this case. Adding a timeout (as done in other places) fix the issue.