improve scope to schema-org types
The meaning of iso19115 scope values are quite poorly documented, most relevant guess seems on comment at http://home.badc.rl.ac.uk/lawrence/blog/2008/03/19/the_scope_of_iso19115

I'm suggesting to use the extension type WebAPI for services, in stead of the dataCatalog type, this is the most relevant change, because service is used quite frequently in the scope of inspire. Following this discussion https://github.com/schemaorg/schemaorg/issues/1423 we may actually be able to link WebAPI's to datasets soon. But we could also wait with merging this PR until schema-org has developed this as part of core