Sign in/out / Add option to sign in in a third party app like GeoServer.
Sometimes users have a GeoServer instance next to GeoNetwork which use
same accounts. When SSO system is not setup, then it requires to sign in
in both apps to access secured layers in the map viewer.

This introduce the possibility to trigger a sign in and sign out action
in a third party before connecting to the catalogue.